Add workaround for system headers on macOS
To to a still unfixed bug in the cmake Xcode generator system header
paths are not included with -isystem but with -I. Therefore if a third
party library header issues warnings this will break the build (because
of -Werror).
This workaround replaces the INTERFACE_INCLUDE_DIRECTORIES property in
cmake with a INTERFACE_COMPILE_OPTIONS directive where -isystem is used
directly. This fixes #10.